Improve example
