+++ src/example/style.scss
@@ -11,3 +11,25 @@
         padding: 0;
     }
 }
+
+.fake-panel {
+    margin: 10px;
+    padding: 5px;
+    box-shadow: 1px 1px 2px 2px #cfcfcf;
+    .head {
+        display: flex;
+        .title {
+            font-weight: bold;
+            font-size: 1.3em;
+            padding: 8px;
+        }
+        .buttons {
+            opacity: 0;
+            flex: 1;
+
+            &:hover {
+                opacity: 1;
+            }
+        }
+    }
+}
